Revert "attacking a time issue with the new graph"
This reverts commit 4279dcef60aa4b8112fccb71ba28661c79a9977a.

+++ client/components/graphing/src/jsMain/kotlin/com/zegreatrob/coupling/client/components/graphing/ScaledTimeFormat.kt
@@ -1,6 +1,6 @@
 package com.zegreatrob.coupling.client.components.graphing
 
-fun scaledTimeFormat(min: Long, max: Long): String {
+fun scaledTimeFormat(min: Int, max: Int): String {
     val range = max - min
     val hasMinutes = (range / (1000 * 60)) > 1
     val hasHours = (range / (1000 * 60 * 60)) > 1
